+++ tests/async/timers.test.tsx
@@ -1,27 +1,31 @@
+import { strictEqual } from "node:assert";
 import test, { suite } from "node:test";
-import { capture, teardown } from "rvx";
+import { capture, Context, teardown } from "rvx";
 import { useAnimation, useInterval, useMicrotask, useTimeout } from "rvx/async";
 import { poll } from "rvx/test";
-import { assertEvents, withMsg } from "../common.js";
-import { strictEqual, throws } from "node:assert";
+import { assertEvents } from "../common.js";
 
 await suite("async/timers", async () => {
+	const CONTEXT = new Context<number>(42);
+
 	await suite("useMicrotask", async () => {
 		await test("regular usage", async () => {
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				useMicrotask(() => {
-					teardown(() => {
-						events.push("teardown");
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					useMicrotask(() => {
+						teardown(() => {
+							events.push("teardown", CONTEXT.current);
+						});
+						events.push("done", CONTEXT.current);
 					});
-					events.push("done");
 				});
 			});
 			assertEvents(events, []);
 			await new Promise<void>(r => queueMicrotask(r));
-			assertEvents(events, ["done"]);
+			assertEvents(events, ["done", 77]);
 			dispose();
-			assertEvents(events, ["teardown"]);
+			assertEvents(events, ["teardown", 42]);
 		});
 
 		await test("immediate disposal", async () => {
@@ -35,17 +39,19 @@ await suite("async/timers", async () => {
 
 		await test("callback disposal", async () => {
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				useMicrotask(() => {
-					teardown(() => {
-						events.push("teardown");
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					useMicrotask(() => {
+						teardown(() => {
+							events.push("teardown", CONTEXT.current);
+						});
+						dispose();
+						events.push("done", CONTEXT.current);
 					});
-					dispose();
-					events.push("done");
 				});
 			});
 			await new Promise<void>(r => queueMicrotask(r));
-			assertEvents(events, ["done", "teardown"]);
+			assertEvents(events, ["done", 77, "teardown", 42]);
 		});
 	});
 
@@ -53,19 +59,21 @@ await suite("async/timers", async () => {
 		await test("regular usage", async () => {
 			const delay = 10;
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				useTimeout(() => {
-					teardown(() => {
-						events.push("teardown");
-					});
-					events.push("done");
-				}, delay);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					useTimeout(() => {
+						teardown(() => {
+							events.push("teardown", CONTEXT.current);
+						});
+						events.push("done", CONTEXT.current);
+					}, delay);
+				});
 			});
 			assertEvents(events, []);
 			await poll(() => events.includes("done"), 1000);
-			assertEvents(events, ["done"]);
+			assertEvents(events, ["done", 77]);
 			dispose();
-			assertEvents(events, ["teardown"]);
+			assertEvents(events, ["teardown", 42]);
 		});
 
 		await test("immediate disposal", async () => {
@@ -81,17 +89,19 @@ await suite("async/timers", async () => {
 		await test("callback disposal", async () => {
 			const delay = 10;
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				useTimeout(() => {
-					teardown(() => {
-						events.push("teardown");
-					});
-					dispose();
-					events.push("done");
-				}, delay);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					useTimeout(() => {
+						teardown(() => {
+							events.push("teardown", CONTEXT.current);
+						});
+						dispose();
+						events.push("done", CONTEXT.current);
+					}, delay);
+				});
 			});
 			await poll(() => events.includes("done"), 1000);
-			assertEvents(events, ["done", "teardown"]);
+			assertEvents(events, ["done", 77, "teardown", 42]);
 		});
 	});
 
@@ -99,22 +109,24 @@ await suite("async/timers", async () => {
 		await test("regular usage", async () => {
 			const delay = 10;
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				let nextId = 0;
-				useInterval(() => {
-					const id = nextId++;
-					teardown(() => {
-						events.push(`e:${id}`);
-					});
-					events.push(`s:${id}`);
-					if (id === 2) {
-						queueMicrotask(dispose);
-						events.push("dispose");
-					}
-				}, delay);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					let nextId = 0;
+					useInterval(() => {
+						const id = nextId++;
+						teardown(() => {
+							events.push(`e:${id}`, CONTEXT.current);
+						});
+						events.push(`s:${id}`, CONTEXT.current);
+						if (id === 2) {
+							queueMicrotask(dispose);
+							events.push("dispose");
+						}
+					}, delay);
+				});
 			});
 			await poll(() => events.includes("dispose"), 1000);
-			assertEvents(events, ["s:0", "e:0", "s:1", "e:1", "s:2", "dispose", "e:2"]);
+			assertEvents(events, ["s:0", 77, "e:0", 42, "s:1", 77, "e:1", 42, "s:2", 77, "dispose", "e:2", 42]);
 			await new Promise<void>(r => setTimeout(r, delay * 2));
 			assertEvents(events, []);
 		});
@@ -132,22 +144,24 @@ await suite("async/timers", async () => {
 		await test("callback disposal", async () => {
 			const delay = 10;
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				let nextId = 0;
-				useInterval(() => {
-					const id = nextId++;
-					teardown(() => {
-						events.push(`e:${id}`);
-					});
-					events.push(`s:${id}`);
-					if (id === 2) {
-						dispose();
-						events.push("dispose");
-					}
-				}, delay);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					let nextId = 0;
+					useInterval(() => {
+						const id = nextId++;
+						teardown(() => {
+							events.push(`e:${id}`, CONTEXT.current);
+						});
+						events.push(`s:${id}`, CONTEXT.current);
+						if (id === 2) {
+							dispose();
+							events.push("dispose");
+						}
+					}, delay);
+				});
 			});
 			await poll(() => events.includes("dispose"), 1000);
-			assertEvents(events, ["s:0", "e:0", "s:1", "e:1", "s:2", "dispose", "e:2"]);
+			assertEvents(events, ["s:0", 77, "e:0", 42, "s:1", 77, "e:1", 42, "s:2", 77, "dispose", "e:2", 42]);
 			await new Promise<void>(r => setTimeout(r, delay * 2));
 			assertEvents(events, []);
 		});
@@ -156,24 +170,26 @@ await suite("async/timers", async () => {
 	await suite("useAnimation", async () => {
 		await test("regular usage", async () => {
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				let nextId = 0;
-				useAnimation(now => {
-					strictEqual(typeof now, "number");
-					const id = nextId++;
-					teardown(() => {
-						events.push(`e:${id}`);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					let nextId = 0;
+					useAnimation(now => {
+						strictEqual(typeof now, "number");
+						const id = nextId++;
+						teardown(() => {
+							events.push(`e:${id}`, CONTEXT.current);
+						});
+						events.push(`s:${id}`, CONTEXT.current);
+						if (id === 2) {
+							queueMicrotask(dispose);
+							events.push("dispose");
+						}
 					});
-					events.push(`s:${id}`);
-					if (id === 2) {
-						queueMicrotask(dispose);
-						events.push("dispose");
-					}
 				});
 			});
 			try {
 				await poll(() => events.includes("dispose"), 1000);
-				assertEvents(events, ["s:0", "e:0", "s:1", "e:1", "s:2", "dispose", "e:2"]);
+				assertEvents(events, ["s:0", 77, "e:0", 42, "s:1", 77, "e:1", 42, "s:2", 77, "dispose", "e:2", 42]);
 			} finally {
 				dispose();
 			}
@@ -193,23 +209,25 @@ await suite("async/timers", async () => {
 
 		await test("callback disposal", async () => {
 			const events: unknown[] = [];
-			const dispose = capture(() => {
-				let nextId = 0;
-				useAnimation(now => {
-					const id = nextId++;
-					teardown(() => {
-						events.push(`e:${id}`);
+			const dispose = CONTEXT.inject(77, () => {
+				return capture(() => {
+					let nextId = 0;
+					useAnimation(now => {
+						const id = nextId++;
+						teardown(() => {
+							events.push(`e:${id}`, CONTEXT.current);
+						});
+						events.push(`s:${id}`, CONTEXT.current);
+						if (id === 2) {
+							dispose();
+							events.push("dispose");
+						}
 					});
-					events.push(`s:${id}`);
-					if (id === 2) {
-						dispose();
-						events.push("dispose");
-					}
 				});
 			});
 			try {
 				await poll(() => events.includes("dispose"), 1000);
-				assertEvents(events, ["s:0", "e:0", "s:1", "e:1", "s:2", "dispose", "e:2"]);
+				assertEvents(events, ["s:0", 77, "e:0", 42, "s:1", 77, "e:1", 42, "s:2", 77, "dispose", "e:2", 42]);
 			} finally {
 				dispose();
 			}